backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pyodbc==5.0.1
Werkzeug==2.3.6
